UploadThings/Repositories/SettingRepositories/ProductRepository.cs
using Microsoft.EntityFrameworkCore;
using UploadThings.Data;
using UploadThings.Models;

namespace UploadThings.Repositories.SettingRepositories
{
    public class ProductRepository : IRepository<Product>
    {
        private readonly MariaDBContext _context;

        public ProductRepository(MariaDBContext context)
        {
            _context = context;
        }

        public async Task<IEnumerable<Product>> GetAllAsync()
        {
            return await _context.Products.ToListAsync();
        }

        public async Task<Product> GetByIdAsync(int id)
        {
            return await _context.Products.FindAsync(id);
        }

        public async Task AddAsync(Product entity)
        {
            await _context.Products.AddAsync(entity);
            await _context.SaveChangesAsync();
        }

        public async Task UpdateAsync(Product entity)
        {
            _context.Products.Update(entity);
            await _context.SaveChangesAsync();
        }

        public async Task DeleteAsync(int id)
        {
            var product = await _context.Products.FindAsync(id);
            if (product != null)
            {
                _context.Products.Remove(product);
                await _context.SaveChangesAsync();
            }
        }
    }

}
